feat: 키보드 > 키보드 그룹으로 변경
